NullMemoryInterface
